Move some probe module stuff to a new python file. Move california pricing tests for 1.6 to a new file.
